refactor folder structure step 1
