+++ src/components/shared/ui/DataPercentage.js
