+++ src/components/shared/ui/DataTranslate.js
