+++ src/components/shared/ui/EmbeddedNumber.js
