+++ src/components/shared/ui/EmbeddedNumberInOut.js
