+++ src/components/shared/ui/Intro.js
